fix: lint issues import/extension
